// src/_services/api/schema/user/user.zod.ts
export const criarUsuarioBody = zod.object({
  "centerId": zod.string(),
  "id": zod.string(),
  "name": zod.string(),
  "primeiroNome": zod.string().optional(),
  "ultimoNome": zod.string().optional(),
  "credencial": zod.string().optional(),
  "turno": zod.enum(['MANHA', 'TARDE', 'NOITE', 'INTERMEDIARIO', 'ADMINISTRATIVO']),
  "processo": zod.string().optional(),
  "empresa": zod.string(),
  "role": zod.enum(['FUNCIONARIO', 'USER', 'ADMIN', 'MASTER']).optional(),
  "roles": zod.union([zod.array(zod.string()),zod.null()]).optional()
})